# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Реализация онлайн библиотеки песен",
        "contact": {
            "name": "Михаил Токмачев",
            "url": "https://t.me/Zatrasz",
            "email": "zatrasz@ya.ru"
        },
        "version": "1.0"
    },
    "host": "localhost:8586",
    "basePath": "/",
    "paths": {
        "/songs": {
            "get": {
                "description": "Принимает поля group , song , releaseDate , text , link .",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Songs"
                ],
                "summary": "Получение данных с фильтрацией по всем полям и пагинацией",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Фильтр по названию",
                        "name": "filter",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "Количество записей на странице для пагинации",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Номер страницы",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Список песен",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Songs"
                            }
                        }
                    },
                    "400": {
                        "description": "Недопустимый параметр смещения",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Ошибка при получение списка песен или Ошибка при обработке запроса",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "description": "Принимает обязательные поля group , song .",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Songs"
                ],
                "summary": "Добавление новой песни",
                "parameters": [
                    {
                        "description": "Данные структуры песни",
                        "name": "songs",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Songs"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешно созданная запись \u003c Ok \u003e",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Неверный формат запроса или не верно заполнены обязательные поля",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Не получены детальные данные из API или ошибка при сохранении в бд",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "description": "Принимает поля songId .",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Songs"
                ],
                "summary": "Удаление песни из библиотеки по ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID записи",
                        "name": "songId",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Данные песни успешно удалены",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Отсутствует идентификатор в запросе",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "не удалось преобразовать строку в число или Ошибка при удалении данных",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "patch": {
                "description": "Принимает поля songId , group , song , releaseDate , text , link . .",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Songs"
                ],
                "summary": "Обновление песни из библиотеки по ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID записи",
                        "name": "songId",
                        "in": "query"
                    },
                    {
                        "description": "Данные структуры песни",
                        "name": "songs",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Songs"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Данные песни успешно обновлены",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Отсутствует идентификатор в запросе",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Не удалось преобразовать строку в число или Ошибка при обновлении данных",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/songs/lyrics": {
            "get": {
                "description": "Принимает поля songId , limit , offset .",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Songs"
                ],
                "summary": "Получение текста песни с пагинацией по куплетам",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID записи",
                        "name": "songId",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "Количество записей на странице для пагинации",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Номер страницы",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Список куплетов",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.LyricResponse"
                            }
                        }
                    },
                    "400": {
                        "description": "Недопустимый параметр смещения или Недопустимый параметр смещения",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Ошибка при получение списка песен или Ошибка при обработке запроса",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "models.LyricResponse": {
            "type": "object",
            "properties": {
                "chunks": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "nextPageID": {
                    "type": "integer"
                }
            }
        },
        "models.SongDetail": {
            "type": "object",
            "properties": {
                "link": {
                    "type": "string"
                },
                "releaseDate": {
                    "type": "string"
                },
                "text": {
                    "type": "string"
                }
            }
        },
        "models.Songs": {
            "type": "object",
            "properties": {
                "ID": {
                    "type": "integer"
                },
                "detail": {
                    "$ref": "#/definitions/models.SongDetail"
                },
                "group": {
                    "type": "string"
                },
                "song": {
                    "type": "string"
                }
            }
        }
    }
}